requirements.txt
torch==2.3.1
torchvision==0.18.1
opencv-python==4.10.0.84
matplotlib==3.9.2
tqdm==4.66.4
pillow==10.4.0
numpy==1.26.4